pyproject.toml
[project]
name = "deep-think-ratio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kernels>=0.12.3",
    "lighteval==0.13.0",
    "lm-eval>=0.4.11",
    "pillow>=11.0.0",
    "ray>=2.54.0",
    "torch",
    "transformers",
    "vllm>=0.18.0",
]

[tool.pyrefly]
search_path = [".", "src"]
